# [stdlib] Eliminate over-constraint on StringProtocol