--- a/validation-test/compiler_crashers_2/0105-sr5050.swift
+++ b/validation-test/compiler_crashers_2/0105-sr5050.swift
@@ -1,8 +1,5 @@
 // RUN: not --crash %target-swift-frontend %s -typecheck
 
-// See <rdar://32555384> 
-// XFAIL: *
-
 protocol P {}
 
 func bar(p: P?) {