@dabrahams 2017-06-04
@swift-ci Please benchmark