@dabrahams 2017-06-04
locally-run benchmark results are as follows.  It seems as though this change degrades performance across the board, which is surprising because almost nothing actually uses or depends on `StringProtocol` conformance.  If these numbers are borne out by the 'bot, I will need help investigating.

<details open>
  <summary>Regression (230)</summary>
  
TEST                                                    | OLD    | NEW    | DELTA  | SPEEDUP 
---                                                     | ---    | ---    | ---    | ---     
ClassArrayGetter                                        | 11     | 15     | +36.4% | **0.73x** 
ArrayPlusEqualFiveElementCollection                     | 3797   | 4700   | +23.8% | **0.81x** 
SuperChars                                              | 140345 | 173508 | +23.6% | **0.81x** 
ArrayOfPOD                                              | 148    | 182    | +23.0% | **0.81x** 
ObjectiveCBridgeStubNSDataAppend                        | 2050   | 2516   | +22.7% | **0.81x** 
ObjectiveCBridgeStubToNSDateRef                         | 3017   | 3681   | +22.0% | **0.82x** 
AnyHashableWithAClass                                   | 60495  | 72829  | +20.4% | **0.83x** 
CaptureProp                                             | 3999   | 4808   | +20.2% | **0.83x** 
ArrayValueProp                                          | 5      | 6      | +20.0% | **0.83x** 
ArrayValueProp4                                         | 5      | 6      | +20.0% | **0.83x** 
ArrayValueProp2                                         | 5      | 6      | +20.0% | **0.83x** 
ArrayValueProp3                                         | 5      | 6      | +20.0% | **0.83x** 
ObjectiveCBridgeStubURLAppendPath                       | 184106 | 218596 | +18.7% | **0.84x** 
NopDeinit                                               | 26372  | 31312  | +18.7% | **0.84x** 
SevenBoom                                               | 1360   | 1613   | +18.6% | **0.84x** 
StackPromo                                              | 21245  | 25193  | +18.6% | **0.84x** 
ObjectiveCBridgeToNSArray                               | 24689  | 29226  | +18.4% | **0.84x** 
SortStringsUnicode                                      | 19407  | 22961  | +18.3% | **0.85x** 
PrefixWhileSequence                                     | 253    | 299    | +18.2% | **0.85x** 
CStringLongAscii                                        | 4241   | 5012   | +18.2% | **0.85x** 
ArrayAppend                                             | 514    | 607    | +18.1% | **0.85x** 
ReversedArray                                           | 50     | 59     | +18.0% | **0.85x** 
RangeAssignment                                         | 262    | 309    | +17.9% | **0.85x** 
ObjectiveCBridgeFromNSDictionaryAnyObject               | 104746 | 123351 | +17.8% | **0.85x** 
PopFrontArrayGeneric                                    | 914    | 1073   | +17.4% | **0.85x** 
ArrayAppendGenericStructs                               | 752    | 881    | +17.2% | **0.85x** 
TwoSum                                                  | 846    | 990    | +17.0% | **0.85x** 
ObjectiveCBridgeStubToNSStringRef                       | 94     | 110    | +17.0% | **0.85x** 
Array2D                                                 | 1940   | 2261   | +16.5% | **0.86x** 
SubstringEquatable                                      | 6598   | 7679   | +16.4% | **0.86x** 
StringWithCString                                       | 103167 | 119850 | +16.2% | **0.86x** 
DictionarySwap                                          | 350    | 406    | +16.0% | **0.86x** 
StringEdits                                             | 94313  | 109392 | +16.0% | **0.86x** 
MapReduce                                               | 276    | 319    | +15.6% | **0.87x** 
DictionaryOfObjects                                     | 2085   | 2409   | +15.5% | **0.87x** 
ObjectiveCBridgeStubNSDateMutationRef                   | 11761  | 13559  | +15.3% | **0.87x** 
DropLastAnySequence                                     | 6261   | 7217   | +15.3% | **0.87x** 
DropLastSequence                                        | 482    | 555    | +15.1% | **0.87x** 
SetExclusiveOr                                          | 2822   | 3246   | +15.0% | **0.87x** 
StringHasSuffixUnicode                                  | 74336  | 85417  | +14.9% | **0.87x** 
StrComplexWalk                                          | 759    | 869    | +14.5% | **0.87x** 
ObjectiveCBridgeStubToNSDate                            | 13475  | 15417  | +14.4% | **0.87x** 
StringFromLongWholeSubstringGeneric                     | 21     | 24     | +14.3% | **0.88x** 
Histogram                                               | 246    | 281    | +14.2% | **0.88x** 
ArrayAppendAscii                                        | 18875  | 21549  | +14.2% | **0.88x** 
StringMatch                                             | 5842   | 6667   | +14.1% | **0.88x** 
UTF8Decode                                              | 220    | 251    | +14.1% | **0.88x** 
SortStrings                                             | 1509   | 1720   | +14.0% | **0.88x** 
StringEqualPointerComparison                            | 251    | 286    | +13.9% | **0.88x** 
ObjectiveCBridgeFromNSArrayAnyObjectToString            | 37687  | 42937  | +13.9% | **0.88x** 
ObjectiveCBridgeStubFromNSDateRef                       | 3209   | 3655   | +13.9% | **0.88x** 
Dictionary2                                             | 1594   | 1815   | +13.9% | **0.88x** 
ArrayAppendLazyMap                                      | 765    | 871    | +13.9% | **0.88x** 
DictionaryLiteral                                       | 1357   | 1544   | +13.8% | **0.88x** 
DictionaryRemove                                        | 2156   | 2447   | +13.5% | **0.88x** 
Join                                                    | 401    | 455    | +13.5% | **0.88x** 
SetUnion_OfObjects                                      | 7137   | 8096   | +13.4% | **0.88x** 
ObjectiveCBridgeFromNSString                            | 1366   | 1549   | +13.4% | **0.88x** 
StringHasPrefixUnicode                                  | 35779  | 40567  | +13.4% | **0.88x** 
IterateData                                             | 675    | 765    | +13.3% | **0.88x** 
PrefixWhileArrayLazy                                    | 45     | 51     | +13.3% | **0.88x** 
DropLastAnySeqCntRangeLazy                              | 15     | 17     | +13.3% | **0.88x** 
SuffixAnySeqCntRangeLazy                                | 15     | 17     | +13.3% | **0.88x** 
ArraySubscript                                          | 1355   | 1535   | +13.3% | **0.88x** 
SetIntersect_OfObjects                                  | 1570   | 1778   | +13.2% | **0.88x** 
PrefixWhileAnySequenceLazy                              | 1445   | 1635   | +13.1% | **0.88x** 
LazilyFilteredArrays                                    | 57735  | 65322  | +13.1% | **0.88x** 
ObjectiveCBridgeFromNSArrayAnyObject                    | 21031  | 23768  | +13.0% | **0.88x** 
SetIntersect                                            | 231    | 261    | +13.0% | **0.89x** 
SuffixAnySequenceLazy                                   | 6164   | 6964   | +13.0% | **0.89x** 
DropWhileAnySequence                                    | 5999   | 6774   | +12.9% | **0.89x** 
ErrorHandling                                           | 2651   | 2993   | +12.9% | **0.89x** 
MapReduceClass                                          | 2849   | 3215   | +12.8% | **0.89x** 
CStringLongNonAscii                                     | 1752   | 1977   | +12.8% | **0.89x** 
ObjectiveCBridgeFromNSStringForced                      | 2010   | 2267   | +12.8% | **0.89x** 
ProtocolDispatch                                        | 2213   | 2495   | +12.7% | **0.89x** 
DropLastAnySeqCRangeIterLazy                            | 4304   | 4852   | +12.7% | **0.89x** 
ObjectiveCBridgeToNSDictionary                          | 43893  | 49453  | +12.7% | **0.89x** 
DropLastSequenceLazy                                    | 490    | 552    | +12.7% | **0.89x** 
ReversedBidirectional                                   | 27191  | 30608  | +12.6% | **0.89x** 
ObjectiveCBridgeFromNSSetAnyObjectToStringForced        | 54942  | 61822  | +12.5% | **0.89x** 
StringWalk                                              | 1000   | 1125   | +12.5% | **0.89x** 
StringUTF16Builder                                      | 1975   | 2221   | +12.5% | **0.89x** 
ArrayOfRef                                              | 3672   | 4125   | +12.3% | **0.89x** 
Dictionary                                              | 487    | 547    | +12.3% | **0.89x** 
ObjectiveCBridgeToNSString                              | 1136   | 1275   | +12.2% | **0.89x** 
PrefixAnyCollectionLazy                                 | 42440  | 47632  | +12.2% | **0.89x** 
DropWhileArray                                          | 33     | 37     | +12.1% | **0.89x** 
ObjectiveCBridgeStubToNSString                          | 1403   | 1573   | +12.1% | **0.89x** 
RGBHistogram                                            | 2123   | 2380   | +12.1% | **0.89x** 
DictionarySwapOfObjects                                 | 6431   | 7209   | +12.1% | **0.89x** 
PrefixAnySequence                                       | 4689   | 5252   | +12.0% | **0.89x** 
PrefixAnySequenceLazy                                   | 4663   | 5219   | +11.9% | **0.89x** 
StringBuilder                                           | 1165   | 1303   | +11.8% | **0.89x** 
SubstringEqualString                                    | 3176   | 3551   | +11.8% | **0.89x** 
StrToInt                                                | 1836   | 2051   | +11.7% | **0.90x** 
ObjectiveCBridgeFromNSSetAnyObjectToString              | 66968  | 74787  | +11.7% | **0.90x** 
LessSubstringSubstringGenericComparable                 | 437    | 488    | +11.7% | **0.90x** 
ObserverUnappliedMethod                                 | 2315   | 2582   | +11.5% | **0.90x** 
ArrayAppendOptionals                                    | 729    | 813    | +11.5% | **0.90x** 
DropWhileAnySeqCntRangeLazy                             | 96     | 107    | +11.5% | **0.90x** 
DropFirstAnyCollectionLazy                              | 43165  | 48098  | +11.4% | **0.90x** 
SuffixAnySeqCRangeIterLazy                              | 4641   | 5171   | +11.4% | **0.90x** 
Dictionary3                                             | 432    | 481    | +11.3% | **0.90x** 
PrefixWhileAnyCollectionLazy                            | 71     | 79     | +11.3% | **0.90x** 
SuffixSequenceLazy                                      | 4207   | 4680   | +11.2% | **0.90x** 
CharacterLiteralsLarge                                  | 5660   | 6296   | +11.2% | **0.90x** 
SuffixAnySeqCRangeIter                                  | 4619   | 5136   | +11.2% | **0.90x** 
PopFrontUnsafePointer                                   | 8027   | 8923   | +11.2% | **0.90x** 
DropFirstSequence                                       | 2025   | 2250   | +11.1% | **0.90x** 
SuffixCountableRange                                    | 9      | 10     | +11.1% | **0.90x** 
ObjectiveCBridgeStubNSDateRefAccess                     | 289    | 321    | +11.1% | **0.90x** 
DropLastAnyCollectionLazy                               | 14461  | 16061  | +11.1% | **0.90x** 
ArrayPlusEqualSingleElementCollection                   | 516    | 573    | +11.0% | **0.90x** 
ProtocolDispatch2                                       | 154    | 171    | +11.0% | **0.90x** 
SuffixAnySequence                                       | 6323   | 7018   | +11.0% | **0.90x** 
ObjectiveCBridgeToNSSet                                 | 37935  | 42102  | +11.0% | **0.90x** 
PrefixAnySeqCRangeIter                                  | 20217  | 22433  | +11.0% | **0.90x** 
HashTest                                                | 1543   | 1712   | +11.0% | **0.90x** 
ArrayOfGenericPOD                                       | 202    | 224    | +10.9% | **0.90x** 
Sim2DArray                                              | 248    | 275    | +10.9% | **0.90x** 
SetUnion                                                | 2563   | 2842   | +10.9% | **0.90x** 
DropLastAnySequenceLazy                                 | 6303   | 6981   | +10.8% | **0.90x** 
PrefixWhileAnySeqCntRange                               | 84     | 93     | +10.7% | **0.90x** 
PrefixWhileCountableRangeLazy                           | 28     | 31     | +10.7% | **0.90x** 
SortSortedStrings                                       | 805    | 891    | +10.7% | **0.90x** 
ArrayAppendRepeatCol                                    | 667    | 738    | +10.6% | **0.90x** 
DropFirstAnySeqCntRangeLazy                             | 47     | 52     | +10.6% | **0.90x** 
EqualSubstringSubstringGenericEquatable                 | 443    | 490    | +10.6% | **0.90x** 
Integrate                                               | 218    | 241    | +10.6% | **0.90x** 
ObjectiveCBridgeFromNSArrayAnyObjectToStringForced      | 33268  | 36769  | +10.5% | **0.90x** 
ObjectiveCBridgeStubURLAppendPathRef                    | 197327 | 217998 | +10.5% | **0.91x** 
PrefixWhileSequenceLazy                                 | 29     | 32     | +10.3% | **0.91x** 
ArrayAppendLatin1                                       | 35340  | 38992  | +10.3% | **0.91x** 
SuffixSequence                                          | 4200   | 4633   | +10.3% | **0.91x** 
DictionaryBridge                                        | 2784   | 3071   | +10.3% | **0.91x** 
MapReduceLazyCollectionShort                            | 39     | 43     | +10.3% | **0.91x** 
EqualSubstringString                                    | 433    | 477    | +10.2% | **0.91x** 
PrefixWhileAnySequence                                  | 13577  | 14954  | +10.1% | **0.91x** 
DropWhileAnySeqCRangeIterLazy                           | 99     | 109    | +10.1% | **0.91x** 
LinkedList                                              | 6592   | 7257   | +10.1% | **0.91x** 
DropLastAnySeqCRangeIter                                | 4370   | 4798   | +9.8%  | **0.91x** 
DropWhileSequenceLazy                                   | 62     | 68     | +9.7%  | **0.91x** 
SortLettersInPlace                                      | 1079   | 1183   | +9.6%  | **0.91x** 
ObjectiveCBridgeStubFromArrayOfNSString                 | 25203  | 27629  | +9.6%  | **0.91x** 
DropWhileArrayLazy                                      | 73     | 80     | +9.6%  | **0.91x** 
StringInterpolation                                     | 10553  | 11564  | +9.6%  | **0.91x** 
SubstringFromLongStringGeneric                          | 157    | 172    | +9.6%  | **0.91x** 
SuffixAnyCollection                                     | 21     | 23     | +9.5%  | **0.91x** 
DropFirstArray                                          | 21     | 23     | +9.5%  | **0.91x** 
DropLastAnyCollection                                   | 21     | 23     | +9.5%  | **0.91x** 
DropFirstArrayLazy                                      | 21     | 23     | +9.5%  | **0.91x** 
Chars                                                   | 758    | 829    | +9.4%  | **0.91x** 
MapReduceClassShort                                     | 4346   | 4751   | +9.3%  | **0.91x** 
PrefixWhileArray                                        | 54     | 59     | +9.3%  | **0.92x** 
DropWhileAnyCollectionLazy                              | 98     | 107    | +9.2%  | **0.92x** 
Hanoi                                                   | 3117   | 3403   | +9.2%  | **0.92x** 
CharacterLiteralsSmall                                  | 307    | 335    | +9.1%  | **0.92x** 
ObjectiveCBridgeStubFromNSDate                          | 3203   | 3495   | +9.1%  | **0.92x** 
ArrayAppendToFromGeneric                                | 373    | 407    | +9.1%  | **0.92x** 
DictionaryRemoveOfObjects                               | 21743  | 23723  | +9.1%  | **0.92x** 
PrefixAnyCollection                                     | 55     | 60     | +9.1%  | **0.92x** 
ObserverForwarderStruct                                 | 1213   | 1323   | +9.1%  | **0.92x** 
ObserverPartiallyAppliedMethod                          | 3559   | 3881   | +9.0%  | **0.92x** 
Prims                                                   | 721    | 786    | +9.0%  | **0.92x** 
ObjectiveCBridgeFromNSSetAnyObjectForced                | 3701   | 4032   | +8.9%  | **0.92x** 
RC4                                                     | 112    | 122    | +8.9%  | **0.92x** 
ObjectiveCBridgeStubDataAppend                          | 3828   | 4169   | +8.9%  | **0.92x** 
ArrayLiteral                                            | 1127   | 1227   | +8.9%  | **0.92x** 
DropWhileAnySequenceLazy                                | 1997   | 2174   | +8.9%  | **0.92x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectToString       | 92089  | 100243 | +8.9%  | **0.92x** 
RecursiveOwnedParameter                                 | 1881   | 2047   | +8.8%  | **0.92x** 
ObjectiveCBridgeFromNSDictionaryAnyObjectToStringForced | 82358  | 89608  | +8.8%  | **0.92x** 
Dictionary2OfObjects                                    | 3113   | 3385   | +8.7%  | **0.92x** 
ObjectiveCBridgeStubFromNSString                        | 790    | 859    | +8.7%  | **0.92x** 
LazilyFilteredRange                                     | 3187   | 3465   | +8.7%  | **0.92x** 
ObjectiveCBridgeFromNSArrayAnyObjectForced              | 4166   | 4529   | +8.7%  | **0.92x** 
PrefixWhileAnySeqCRangeIter                             | 12101  | 13150  | +8.7%  | **0.92x** 
EqualSubstringSubstring                                 | 445    | 483    | +8.5%  | **0.92x** 
PrefixAnySeqCRangeIterLazy                              | 20523  | 22261  | +8.5%  | **0.92x** 
DropWhileAnySeqCRangeIter                               | 20496  | 22229  | +8.5%  | **0.92x** 
ArrayAppendFromGeneric                                  | 379    | 411    | +8.4%  | **0.92x** 
PrefixAnySeqCntRangeLazy                                | 48     | 52     | +8.3%  | **0.92x** 
Phonebook                                               | 5608   | 6074   | +8.3%  | **0.92x** 
ObjectiveCBridgeStubDateAccess                          | 170    | 184    | +8.2%  | **0.92x** 
DropWhileSequence                                       | 1533   | 1659   | +8.2%  | **0.92x** 
DropWhileAnySeqCntRange                                 | 61     | 66     | +8.2%  | **0.92x** 
PopFrontArray                                           | 965    | 1043   | +8.1%  | **0.93x** 
PrefixWhileCountableRange                               | 25     | 27     | +8.0%  | **0.93x** 
DropFirstAnySeqCRangeIterLazy                           | 25914  | 27972  | +7.9%  | **0.93x** 
Walsh                                                   | 315    | 340    | +7.9%  | **0.93x** 
SubstringComparable                                     | 3525   | 3797   | +7.7%  | **0.93x** 
DeadArray                                               | 143    | 154    | +7.7%  | **0.93x** 
DropWhileCountableRangeLazy                             | 65     | 70     | +7.7%  | **0.93x** 
PrefixSequenceLazy                                      | 1511   | 1627   | +7.7%  | **0.93x** 
ObjectiveCBridgeStubFromNSStringRef                     | 144    | 155    | +7.6%  | **0.93x** 
ArrayAppendUTF16                                        | 33167  | 35691  | +7.6%  | **0.93x** 
DropFirstAnySequence                                    | 5797   | 6228   | +7.4%  | **0.93x** 
MonteCarloE                                             | 9576   | 10281  | +7.4%  | **0.93x** 
StringFromLongWholeSubstring                            | 9249   | 9924   | +7.3%  | **0.93x** 
…